Adds set up code
A couple of common components have been created:

- Button
- Container

The Words file just contains an array of strings, words that will be typed in-game

Finally, in App we have set up several slices of state (currently commented out) that will be used in the making of this game. In this approach all the state is living in the top layer - which is App.js.